pyproject.toml
[project]
name = "appzaralat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiogram>=3.25.0",
    "python-dotenv>=1.2.1",
]
